fix(ci): add IDesignTimeDbContextFactory so migrate.yml doesn't need to boot the whole app
Found via live testing: dotnet ef has no lightweight path to a DbContext
for a top-level-statement Program.cs, so it was booting the ENTIRE app
pipeline just to run database update. That needed a pile of unrelated
config (JWT key, email creds, ALLOWED_ORIGINS) the migration job has no
business setting — and worse, would have tripped this same PR's new
startup fail-fast check (pending migrations -> throw), since CI doesn't
set ASPNETCORE_ENVIRONMENT=Development. A migration job whose only way to
run needs the exact check it's supposed to satisfy already passing is a
non-starter.

ApplicationDbContextFactory : IDesignTimeDbContextFactory<ApplicationDbContext>
is the standard EF Core answer — dotnet ef prefers it over booting
Program.cs when present, and it only needs
ConnectionStrings__POSTGRES_CONNECTION_STRING. Extracted the postgres://
-> Npgsql key=value conversion (previously a local function inside
Program.cs) into Server/Infrastructure/PostgresConnectionString.cs so
both the real app and the design-time factory share one implementation.

migrate.yml simplified to match: sets the env var directly instead of
piping it through --connection (which the factory wouldn't read anyway
since it reads the environment, not CLI args).

Verified locally with `env -i` (completely clean environment, only
ConnectionStrings__POSTGRES_CONNECTION_STRING set) — dotnet ef database
update succeeds with zero other config present, proving the app
pipeline is genuinely bypassed.

dotnet test 502/502.
