+++ Server/Program.cs
@@ -33,32 +33,7 @@ var rawConnectionString = builder.Configuration.GetConnectionString("POSTGRES_CO
                           throw new InvalidOperationException("Connection string 'POSTGRES_CONNECTION_STRING' not found.");
 
 // Support both postgres:// URL format and Npgsql key=value format
-var connectionString = rawConnectionString.StartsWith("postgres://") || rawConnectionString.StartsWith("postgresql://")
-    ? ConvertPostgresUrl(rawConnectionString)
-    : rawConnectionString;
-
-static string ConvertPostgresUrl(string url)
-{
-    var uri = new Uri(url);
-    var userInfo = uri.UserInfo.Split(':');
-    var username = userInfo[0];
-    var password = userInfo.Length > 1 ? Uri.UnescapeDataString(userInfo[1]) : string.Empty;
-    var host = uri.Host;
-    var port = uri.Port > 0 ? uri.Port : 5432;
-    var database = uri.AbsolutePath.TrimStart('/');
-    var query = System.Web.HttpUtility.ParseQueryString(uri.Query);
-    var sslMode = query["sslmode"] ?? "Prefer";
-    var npgsqlSsl = sslMode.ToLowerInvariant() switch {
-        "require"     => "Require",
-        "verify-ca"   => "VerifyCA",
-        "verify-full" => "VerifyFull",
-        "disable"     => "Disable",
-        _             => "Prefer"
-    };
-    // Only trust-cert for Require/Prefer — VerifyCA/VerifyFull must validate the chain
-    var trustCert = npgsqlSsl is "Require" or "Prefer" ? ";Trust Server Certificate=true" : string.Empty;
-    return $"Host={host};Port={port};Username={username};Password={password};Database={database};SSL Mode={npgsqlSsl}{trustCert}";
-}
+var connectionString = FourPlayWebApp.Server.Infrastructure.PostgresConnectionString.Normalize(rawConnectionString);
 builder.Services.AddOptions();
 builder.Services.AddControllersWithViews();
 #region Email
